+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Tests for Credfeto.Notification.Bot.Twitch.Models to achieve 100% code coverage
 - Added unit test project for Credfeto.Notification.Bot.Server achieving 100% code coverage
 - Tests for Credfeto.Notification.Bot.Twitch.DataTypes assembly - Streamer, Viewer, TwitchUser, and mapper types
+- Increased code coverage for Credfeto.Notification.Bot.Twitch to 100%
 ### Fixed
 ### Changed
 - Dependencies - Updated Credfeto.Enumeration to 1.2.144.1906